Normalise gorailka to Ukranian
This is what we had in _With Fire and Sword_, but I changed it for some reason in this production. It doesn’t seem to be modern Ukranian, but everything I can find lists this word to be of Ukranian origin so let’s standardise on this for the time being.